[WHAL-58] Refactor: Vote, Choice, Coin serializer 분리 및 패키지화
