+++ vote/serializers/coin_serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from vote.models import Coin
+
+
+class CoinSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Coin
+        exclude = ['coin_krname']
