+++ vote/views.py
@@ -2,17 +2,17 @@ from django.shortcuts import get_object_or_404
 from rest_framework import viewsets, mixins
 from rest_framework.response import Response
 
-from .serializers import VoteCreateSerializer, VoteListSerializer, VoteDetailSerializer, ChoiceSerializer
+from .serializers import coin_serializers, vote_serializers, choice_serializers
 from .models import Vote, Choice
 
 
 class VoteViewSet(viewsets.GenericViewSet, mixins.CreateModelMixin):
     queryset = Vote.objects.all()
-    serializer_class = VoteCreateSerializer
+    serializer_class = vote_serializers.VoteCreateSerializer
 
     def list(self, request):
         queryset = self.get_queryset()
-        serializer = VoteListSerializer(queryset, many=True)
+        serializer = vote_serializers.VoteListSerializer(queryset, many=True)
         for data in serializer.data:
             participants = data.pop('participants')
             if request.user.user_id in participants:
@@ -25,7 +25,7 @@ class VoteViewSet(viewsets.GenericViewSet, mixins.CreateModelMixin):
     def retrieve(self, request, pk=None):
         queryset = self.get_queryset()
         vote = get_object_or_404(queryset, pk=pk)
-        serializer = VoteDetailSerializer(vote, context={'user': request.user})
+        serializer = vote_serializers.VoteDetailSerializer(vote, context={'user': request.user})
         return Response(serializer.data)
 
 
@@ -35,5 +35,5 @@ class VoteViewSet(viewsets.GenericViewSet, mixins.CreateModelMixin):
 
 
 class ChoiceViewSet(viewsets.ModelViewSet):
-    serializer_class = ChoiceSerializer
+    serializer_class = choice_serializers.ChoiceSerializer
     queryset = Choice.objects.all()
